pyproject.toml
[tool.poetry]
name = "trading-app"
version = "0.1.0"
description = ""
authors = ["pampagrimaldi <pampagrimaldi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
black = "^23.11.0"
alembic = "^1.12.1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
alpaca-trade-api = "^3.0.2"
polygon-api-client = "^1.13.1"
tqdm = "^4.66.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
lean = "^1.0.174"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
